requirements.txt
boto3
botocore
Flask
Flask_Limiter
Flask_Migrate
flask_sqlalchemy
flask-talisman
flask-inputs
geopy
python-dotenv
pytz
qrcode
redis
Requests
sentry_sdk
SQLAlchemy
watchtower
Werkzeug
pillow
gunicorn